+++ test/csv_test.rb
@@ -78,6 +78,11 @@ class CsvTest < Minitest::Test
     assert_equal "a,b\n1,one\n2,two\n3,three\n", df.write_csv
   end
 
+  def test_to_csv
+    df = Polars::DataFrame.new({"a" => [1, 2, 3], "b" => ["one", "two", "three"]})
+    assert_equal "a,b\n1,one\n2,two\n3,three\n", df.to_csv
+  end
+
   def test_has_header_true
     df = Polars.read_csv("test/support/data.csv", has_header: true)
     assert_equal ["a", "b"], df.columns